+++ AddOns/Quartz/CHANGELOG.md
@@ -1,6 +1,12 @@
 # Quartz
 
-## [3.7.15a](https://github.com/Nevcairiel/Quartz/tree/3.7.15a) (2025-07-03)
-[Full Changelog](https://github.com/Nevcairiel/Quartz/compare/3.7.15...3.7.15a) [Previous Releases](https://github.com/Nevcairiel/Quartz/releases)
+## [3.7.16](https://github.com/Nevcairiel/Quartz/tree/3.7.16) (2025-03-04)
+[Full Changelog](https://github.com/Nevcairiel/Quartz/compare/3.7.15a...3.7.16) [Previous Releases](https://github.com/Nevcairiel/Quartz/releases)
 
-- Split TOC into different lines again  
+- Update retail channeling ticks  
+    Add channeling ticks for several spells.  
+    Remove Essence Font and unused versions of Void Torrent and Shifting Power.  
+- Update Channeling Ticks for Mists of Pandaria Classic  
+    Based on data provided in #99  
+    Fixes #99  
+    Fixes #100